Remove libc
Not everything could be remove without breaking the existing API. In particular, `ts_tree_print_dot_graph` function takes a `FILE *` as input parameter.

This function has been moved to a separate file called `capi.rs`, which is compiled when the `capi` feature is available. Internally the project uses a `BufWriter<File>` normally, but when `capi` feature is requested, a `CFile` is stored. This is an abstraction layer between `FILE *` and `io::Write`, allowing the usage of C stream functions in a transparent way.

+++ tree-sitter-tests/src/bindings.rs
@@ -599,7 +599,7 @@ impl Parser {
         if let Some(flag) = flag {
             ffi::ts_parser_set_cancellation_flag(
                 self.0.as_ptr(),
-                flag as *const AtomicUsize as *const u64,
+                flag as *const AtomicUsize as *const usize,
             );
         } else {
             ffi::ts_parser_set_cancellation_flag(self.0.as_ptr(), ptr::null());
